requirements.txt
PIL
Pycryptodome
bcrypt==3.1.7
PyMySQL==0.9.3
